requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
python-multipart
python-dotenv
